Update index.html
Updated index.html:
Cleaned up design by adding background banner for Destinations and added JMU colors to it and to the select locations or button.
Added a button that for Additional Location Information.
When pressed the a popup appears With the full names of the locations and hyperlinks to the websites they correspond to.