Added Tour component for User Tutorial.
It was very hard to add the tour without adding styling within the tsx I may have to find a way to clean up the code later. I also wanted to add a step to show the stop button but may be hard to do if it only shows during navigation.